+++ src/Model/Command/Sorting.php
@@ -14,6 +14,8 @@ class Sorting extends AbstractCommand implements UserAwareInterface
     private $userId;
     /** @var string[] */
     private $itemIds = [];
+    /** @var string|null */
+    private $modelName = null;
 
     private function __construct(string $userId, array $itemIds)
     {
@@ -29,6 +31,20 @@ class Sorting extends AbstractCommand implements UserAwareInterface
     public static function create(string $userId, array $itemIds): self
     {
         return new static($userId, $itemIds);
+    }
+
+    /**
+     * Set A/B model name
+     *
+     * @return $this
+     */
+    public function setModelName(string $modelName): self
+    {
+        Assertion::typeIdentifier($modelName);
+
+        $this->modelName = $modelName;
+
+        return $this;
     }
 
     public function getUserId(): string
@@ -57,9 +73,15 @@ class Sorting extends AbstractCommand implements UserAwareInterface
 
     protected function getCommandParameters(): array
     {
-        return [
+        $parameters = [
             'user_id' => $this->userId,
             'item_ids' => $this->itemIds,
         ];
+
+        if ($this->modelName !== null) {
+            $parameters['model_name'] = $this->modelName;
+        }
+
+        return $parameters;
     }
 }