+++ tests/unit/Model/Command/SortingTest.php
@@ -11,9 +11,13 @@ class SortingTest extends UnitTestCase
     {
         $userId = 'user-id';
         $itemIds = ['item-1', 'item-3', 'item-2'];
+        $modelName = 'test-model-name';
 
         $command = Sorting::create($userId, $itemIds);
         $this->assertSortingCommand($command, $userId, $itemIds);
+
+        $command->setModelName($modelName);
+        $this->assertSortingCommand($command, $userId, $itemIds, $modelName);
     }
 
     /**
@@ -21,16 +25,22 @@ class SortingTest extends UnitTestCase
      *
      * @param Sorting $command
      */
-    private function assertSortingCommand($command, string $userId, array $itemIds): void
+    private function assertSortingCommand($command, string $userId, array $itemIds, ?string $modelName = null): void
     {
+        $parameters = [
+            'user_id' => $userId,
+            'item_ids' => $itemIds,
+        ];
+
+        if ($modelName !== null) {
+            $parameters['model_name'] = $modelName;
+        }
+
         $this->assertInstanceOf(Sorting::class, $command);
         $this->assertSame(
             [
                 'type' => 'sorting',
-                'parameters' => [
-                    'user_id' => $userId,
-                    'item_ids' => $itemIds,
-                ],
+                'parameters' => $parameters,
             ],
             $command->jsonSerialize()
         );